% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd_phi.R
\name{elpd_phi}
\alias{elpd_phi}
\title{The ELPD as a function of phi and eta}
\usage{
elpd_phi(x, eta, const)
}
\arguments{
\item{x}{the value of phi}

\item{eta}{learning rate}

\item{const}{constants}
}
\value{
numeric
}
\description{
The ELPD as a function of phi and eta
}
