% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd_phi_2.R
\name{elpd_phi_2}
\alias{elpd_phi_2}
\title{Evaluates the ELPD of phi given eta, x1, and x2}
\usage{
elpd_phi_2(eta, x1, x2, const)
}
\arguments{
\item{eta}{the learning rate}

\item{x1}{the sample mean of the first dataset}

\item{x2}{the sample mean of the second dataset}

\item{const}{experiment constant}
}
\value{
numeric
}
\description{
Evaluates the ELPD of phi given eta, x1, and x2
}
