% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd_phi_theta.R
\name{elpd_phi_theta}
\alias{elpd_phi_theta}
\title{Calculate the ELPD of phi + theta}
\usage{
elpd_phi_theta(x, eta, const)
}
\arguments{
\item{x}{the value of phi + theta}

\item{eta}{learning rate}

\item{const}{constants}
}
\value{
ELPD value
}
\description{
Calculate the ELPD of phi + theta
}
