% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_eta.R
\name{lambda_eta}
\alias{lambda_eta}
\title{Return the value of lambda for a specific eta}
\usage{
lambda_eta(eta, const)
}
\arguments{
\item{eta}{learning rate}

\item{const}{a list of contants}
}
\value{
a vector
}
\description{
Return the value of lambda for a specific eta
}
