% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_mean.R
\name{phi_mean}
\alias{phi_mean}
\title{Posterior phi given eta, x1, and x2}
\usage{
phi_mean(eta, x1, x2, const)
}
\arguments{
\item{eta}{learning rate}

\item{x1}{sample average for the first dataset}

\item{x2}{sample average for the second dataset}

\item{const}{constants}
}
\value{
posterior phi
}
\description{
Posterior phi given eta, x1, and x2
}
