% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_theta_mean.R
\name{phi_theta_mean}
\alias{phi_theta_mean}
\title{Posterior mean of phi + theta}
\usage{
phi_theta_mean(eta, x1, x2, const)
}
\arguments{
\item{eta}{learning rate}

\item{x1}{sample mean of the first dataset}

\item{x2}{sample mean of the second dataset}

\item{const}{constants}
}
\description{
Posterior mean of phi + theta
}
