% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_theta_var.R
\name{phi_theta_var}
\alias{phi_theta_var}
\title{Posterior variance of phi + theta}
\usage{
phi_theta_var(eta, const)
}
\arguments{
\item{eta}{learning rate}

\item{const}{constants}

\item{x1}{sample mean of the first dataset}

\item{x2}{sample mean of the second dataset}
}
\description{
Posterior variance of phi + theta
}
