% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elpd_phi.R
\name{plot_elpd_phi}
\alias{plot_elpd_phi}
\title{Plot the envelope of the ELPD for the first dataset}
\usage{
plot_elpd_phi(xmin, xmax, x1, x2)
}
\arguments{
\item{xmin}{the minimum value of the x-axis}

\item{xmax}{the maximum value of the x-axis}

\item{x1}{the sample mean of the first dataset}

\item{x2}{the sample mean of the second dataset}
}
\value{
ggplot object
}
\description{
Plot the envelope of the ELPD for the first dataset
}
