% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elpd_phi_eta.R
\name{plot_elpd_phi_eta}
\alias{plot_elpd_phi_eta}
\title{Plot the ELPD as a function of eta}
\usage{
plot_elpd_phi_eta(x1, x2)
}
\arguments{
\item{x1}{the sample mean of the first dataset}

\item{x2}{the sample mean of the second dataset}
}
\value{
ggplot object
}
\description{
Plot the ELPD as a function of eta
}
