% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elpd_phi_theta.R
\name{plot_elpd_phi_theta}
\alias{plot_elpd_phi_theta}
\title{Plot the ELPD for $\hat\varphi_\eta + \theta | \hat\varphi_\eta$}
\usage{
plot_elpd_phi_theta(xmin, xmax, x1, x2)
}
\arguments{
\item{xmin}{minimum value of the x-axis}

\item{xmax}{maximum value of the x-axis}

\item{x1}{sample mean of the first dataset}

\item{x2}{sample mean of the second dataset}
}
\value{
ggplot object
}
\description{
Plot the ELPD for $\hat\varphi_\eta + \theta | \hat\varphi_\eta$
}
