% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_mean.R
\name{theta_mean}
\alias{theta_mean}
\title{Conditional posterior mean of theta given phi}
\usage{
theta_mean(eta, x1, x2, const)
}
\arguments{
\item{eta}{learning rate}

\item{x1}{the sample mean of the first dataset}

\item{x2}{the sample mean of the second dataset}

\item{const}{constants}
}
\description{
Conditional posterior mean of theta given phi
}
